requirements.txt
Flask==3.1.0
google_api_python_client==2.161.0
mysql-connector-python==8.0.33
protobuf==3.20.3
Werkzeug==3.1.3
pymongo
python-dotenv